+++ labrad_server/open_scalabrad.bat
@@ -1,1 +1,1 @@
-M:\scalabrad\bin\labrad --registry file:///M:/Registry?format=delphi
+M:\scalabrad-0.8.3\bin\labrad --registry file:///M:/Registry?format=delphi
