--- zi_config.ini
@@ -1,4 +1,0 @@
-[config]
-_default_dir = M:\Experimental Data
-_default_session = ['', 'hwh', '20200930_12Q_sample_C_test']
-
